Inserção de imagem de user completa
Falta o uso do JCrop para cortar a imagem

+++ DAL/Perfil/Class_inserirUfoto.php
@@ -1,16 +1,26 @@
 <?php
-    try
-    {
-        $UimgNNo = $_SESSION['User_Name'] . $UimgExt; //Nomeia o arquivo
-        $UimgPasta = "../../uploads/img_Uperf"; //Defini o nome da pasta em que o arquivo será salvo
-                    
-        move_uploaded_file($_FILES['Uimg']['tmp_name'], $UimgPasta.$UimgNNo);//efetua o upload
+    //Incluir a conexão com banco de dados
+    include_once '../../nav_home.php';
 
-        $_SESSION['UimgNNo'] = $UimgNNo; //Amazena o novo nome na sessão
-    }
-    catch(Exception $ex)
+
+    $UimgExt = strtolower(substr($_FILES['Uimg']['name'],-4));//Pega o nome da extesão
+    $UimgNNo = $_SESSION['User_Name'] . $UimgExt; //Nomeia o arquivo
+    $UimgPasta = "../../uploads/img_Uperf/"; //Defini o nome da pasta em que o arquivo será salvo
+    
+    if(file_exists($UimgPasta.$UimgNNo))
     {
-        $_SESSION['UimgNNo'] = null;
+        unlink($UimgPasta.$UimgNNo); //Deleta caso arquivo já exista
+        move_uploaded_file($_FILES['Uimg']['tmp_name'], $UimgPasta.$UimgNNo);//efetua o upload
     }
-    header("Location: Class_alterarU_DAL.php");
+    else
+    {
+        move_uploaded_file($_FILES['Uimg']['tmp_name'], $UimgPasta.$UimgNNo);//efetua o upload
+    }
+
+    $_SESSION['UimgNNo'] = $UimgNNo; //Amazena o novo nome na sessão
+
+    $_SESSION['auxiliar'] = "Sucesso ao alterar seus dados"; 
+
+    header("Location: ../../perfil.php");
+
 ?>
